// src/utils/enums/error-code-name.enums.ts
export enum ErrorTypes {
  TOKEN_EXPIRED = 'TOKEN_EXPIRED',
  INVALID_TOKEN = 'INVALID_TOKEN',
  ACCOUNT_INACTIVE = 'ACCOUNT_INACTIVE',
  REFRESH_EXPIRED = 'REFRESH_EXPIRED',
  IOS_CRITICAL_UPDATE = 'IOS_CRITICAL_UPDATE',
  ANDROID_CRITICAL_UPDATE = 'ANDROID_CRITICAL_UPDATE',
  APPLICATION_MAINTENANCE = 'APPLICATION_MAINTENANCE',
  APPLICATION_STOPPED = 'APPLICATION_STOPPED',
  PASSWORD_NOT_VALID = 'PASSWORD_NOT_VALID',
  USER_BLOCKED = 'USER_BLOCKED',
  USER_DELETED = 'USER_DELETED',
  SUSPENDED = 'SUSPENDED',
  OTP_NOT_MATCH = 'OTP_NOT_MATCH',
  OTP_NOT_EXIST_OR_EXPIRED = 'OTP_NOT_EXIST_OR_EXPIRED',
  ACCESS = 'ACCESS',
  NOT_FOUND = 'NOT_FOUND',
  NOT_FOUND_PRODUCT = 'NOT_FOUND_PRODUCT',
  EMAIL_EXIST = 'EMAIL_EXIST',
  PHONE_EXIST = 'PHONE_EXIST',
  NOT_FOUND_USER = 'NOT_FOUND_USER',
  MANY_ATTEMPT = 'MANY_ATTEMPT',
  PERMISSION = 'PERMISSION',
  NOT_COMPLETED = 'NOT_COMPLETED',
  DUPLICATED = 'DUPLICATED',
  ERROR_UPDATE = 'ERROR_UPDATE',
  PENDING_REQUESTS = 'PENDING_REQUESTS',
  INVALID_URL_PARAMS = 'INVALID_URL_PARAMS',
  CLOSED_ORDER = 'CLOSED_ORDER',
  TOKEN_REQUIRED = 'TOKEN_REQUIRED',
  NOT_ALLOWED = 'NOT_ALLOWED',
  USER_GOV_DIS_NOT_FOUND = 'USER_GOV_DIS_NOT_FOUND',
  CSRF_TOKEN_REQUIRED = 'CSRF_TOKEN_REQUIRED',
  CSRF_INVALID_TOKEN = 'CSRF_INVALID_TOKEN',
  CSRF_TOKEN_EXPIRED = 'CSRF_TOKEN_EXPIRED',
  WRONG_CSRF_TOKEN = 'WRONG_CSRF_TOKEN',
}